change folder structure to features
