--- client/src/constants/action-types.js
@@ -1,5 +1,0 @@
-
-export const ACTION_TYPES = {
-  thisAction: 'THIS_ACTION',
-  thatAction: 'THAT_ACTION'
-};
