+++ client/src/features/app/app.types.js
@@ -1,0 +1,4 @@
+export const APP_TYPES = {
+    thisAction: 'THIS_ACTION',
+    thatAction: 'THAT_ACTION'
+  };
