+++ client/src/features/list-item/list-item.types.js
@@ -1,0 +1,4 @@
+export const ACTION_TYPES = {
+    thisAction: 'THIS_ACTION',
+    thatAction: 'THAT_ACTION'
+  };
