+++ client/src/reducers.js
@@ -1,0 +1,7 @@
+import { combineReducers } from 'redux';
+import { listItemReducer } from "./features/list-item/list-item.reducers";
+
+
+export default combineReducers({
+	listItemReducer
+});
