--- client/src/reducers/index.js
@@ -1,7 +1,0 @@
-import { combineReducers } from 'redux';
-import {sampleReducer} from "./sample-reducer";
-
-
-export default combineReducers({
-	sampleReducer
-});
